fix all register controller bugs

+++ src/utils/cloudinary.js
@@ -1,7 +1,7 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
-cloudinary.v2.config({
+cloudinary.config({
   cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
   api_key: process.env.CLOUDINARY_API_KEY,
   api_secret: process.env.CLOUDINARY_API_SECRET,
@@ -22,10 +22,16 @@ const uploadOnCloudinary = async (localFilePath) => {
     });
 
     // file uploaded successfully
-    console.log("file uploaded successfully on cloudinary", response.url);
+    // console.log("file uploaded successfully on cloudinary", response.url);
+    // console.log("Response: ", response);
+
+    fs.unlinkSync(localFilePath); // Removes temp file
+
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath); // Removes temp file
+    if (localFilePath) {
+      fs.unlinkSync(localFilePath); // Removes temp file
+    }
     return null;
   }
 };
